Updated ListAllInteractor + Associated files
Updated the ListAllInteractor to accept the checkForUpdatesBoundary as part of the constructor instead of as a parameter to ListAll.  This should hopefully simplify calls to ListAllInteractor and eliminate the need to pass in a updateChecker object, making it the CommandsFactory's responsibility instead.

Modified CommandsFactory, ListController, ListAllBoundary to accept this new presentation method.