Created Presenter for the ListAll usecase
Instead of logging directly in the Controller, the implementation of the presenter allows for logging to take place in the ListPresenter instead (as it should).

Implementation Details: ListController now instantiates a new ListPresenter based on the log parameter, which performs the logging task with error codes, etc.

ListResult is a record class storing the result of a call to ListAll in the controller.

ListResultPresenter offers an interface for the ListPresenter's displayResult method, as pursuant to the Dependency Inversion Principle.